front_end_app/requirements.txt
Flask==3.1.1
oracledb==3.2.0
pandas==2.3.1
pdf2image==1.17.0
Pillow==11.3.0
protobuf==6.31.1
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
Requests==2.32.4
SQLAlchemy==2.0.41
streamlit==1.45.0
streamlit_lottie==0.0.5
Werkzeug==3.1.3
